NVMe: Implement Controller Memory Buffer
Specification 1.2 provides a new feature for controller to provide memory
through a BAR that can be used by the host for various reasons.

With this commit, the memory is provdied through an exclusive BAR and
hardcoded to 128MB. It can only be used for CQ and SQ entries as well
as PRP lists. Read/Write data is not supported

Signed-off-by: Keith Busch <keith.busch@intel.com>