Enhancements to the CMB feature for NVM
This patch adds additional configurability to the Controller Memory
Buffer of the QEMU 1.2 device and adds support for the CMBLOC and
CMBSZ registers which were added as part of the 1.2 specification.

The command-line argument cmb=<cmb> is now more than a flag and can be
used to set the size of the CMB (in MB). In additional the CMBLOC and
CMBSZ register are setup to reflect their appropriate values.

Signed-off-by: Stephen Bates <sbates@raithlin.com>
Signed-off-by: Keith Busch <keith.busch@intel.com>