lightnvm: expose several LUNs to the host
Allow the LightNVM nvme device to expose several LUNs to the host. This
brings as a consequence that the physical address space cannot be
assumed sequential anymore. We use the ppa list to both, send the right
addresses to the dma handler and to update the translation table
accordingly.

Also, since the LightNVM and standard NVMe paths have diverged
substantially, separate them from the moment the command arrives.