dma: allow r/w dma handlers to use addr list
In the general case, addresses are assumed sequential, and a starting
lba address is sent down the stack to store and retrieve data. In case
that the device driver sends a list of addresses to write to / read
from, enable dma functions that take the addresses from a list, instead
of incrementing the starting address.