NVMe: lnvm: Added "lnvm_init_meta"
Added initialization of meta-data by creating a meta-data file with state
"UNKNOWN" for all sectors.

Prior to this change, then state checks did not properly recognize state when
meta-data file is empty, it was possible to do writes to blocks that were not
erased, do reads from files that were not written, etc.