nvme: refactor block backend i/o
This is a major refactor of how I/O to the block backend is handled. The
QEMUSGLists and QEMUIOVectors produced by nvme_map_prp are now sent
through a new path where they may be split into QSGs/IOVs for the block
backend and one QSG/IOV for data that is predefined.

For the vector path, the patch eliminates the need for the list_io
functions in dma-helpers.c by instead submitting multiple requests. This
has the added benefit of supporting PRP lists where the entries are not
necessarily aligned to a page/sector boundary.

This also moves metadata to the block backend.

Signed-off-by: Klaus B. A. Jensen <klaus.jensen@cnexlabs.com>