nvme: expand DLFEAT support
The new `dlfeat` NVMe device parameter can now be used to define what
predefined data should be used. If set to 0x0 no data will be copied
back to the host.

A single LBADS-sized block is used to hold the predefined data and I/O
to LBAs that should result in predefined data are mapped to this.

Signed-off-by: Klaus B. A. Jensen <klaus.jensen@cnexlabs.com>