nvme: clean up multi namespace support
This commit cleans up the support for multiple namespaces. Currently
most parameters affecting namespaces (such as DLFEAT) is still common to
all created namespaces.

Remove the `nlbaf` and `lba_index` configuration parameters in favor of
the `ms` and the new `ms_max` parameter. The device will now generate
LBA formats with a fixed data size according to the `sec_size`
parameter and produce formats with a meta data size ranging from 0 to
`ms_max` in power of two increments. The `ms` parameter controls what
format will be selected by default.

Also, in the interest of getting this upstreamed, start to support the
notion of "dialects" such that the device can present itself as both a
regular NVMe 1.3 device as well as an OpenChannel device. This is work
in progress but this commit sketches a possible method of achieving
this.

Signed-off-by: Klaus B. A. Jensen <klaus.jensen@cnexlabs.com>