nvme: fix performance regression
Expanding the log page size to 4 MB had the side effect of severely
slowing down writes due to the device writing the entire log page to
backend on each write (when advancing the write pointer).

This commit fixes that such that updates to chunk state is written
individually on a per chunk basis.

Signed-off-by: Klaus B. A. Jensen <klaus.jensen@cnexlabs.com>