nvme: allow specifying number of chunks per punit
Previously the device would deduce the number of chunks from the
`lnum_grp`, `lnum_pu` and `lnum_sec` QEMU device parameters and the size
of the underlying device. This continues to be the default behavior.

Optionally, one can now use the `lnum_chk` device parameter to
explicitly define the geometry. The device will make sure that the
underlying device and chunk info log page size can accomodate the
requested geometry.

Signed-off-by: Klaus B. A. Jensen <klaus.jensen@cnexlabs.com>