These changes adds a new 'ocssd' block driver.
The block format is basically a raw image and can still be used as such.
The effect of adding this new block driver is that an initialized image
file can be created using qemu-img:

    qemu-img create -f ocssd -o num_grp=2,num_pu=4,num_chk=60 \
        ocssd.img

There are other options availabe, see `-o help`. The command will
initialize an image with a header and a copy of the OCSSD Geometry Data
Structure describing the geometry and write characteristics (WS_MIN,
etc.) and automatically calculate the needed image size.

This also means that geometry related parameters have been removed from
the nvme device (`num_grp` and friends). Parameters to control write
characteristics (`ws_min` and friends) are still available.

Signed-off-by: Klaus B. A. Jensen <klaus.jensen@cnexlabs.com>